DOCS/ubuai_server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
websockets==12.0
requests==2.31.0
python-dotenv==1.0.1
pyttsx3==2.90
pydub==0.25.1
